getAll chs method created

+++ backend/src/alek/model/ch.go
@@ -2,7 +2,7 @@ package model
 
 type Ch struct {
 	Id          string   `json:"id"`
-	AvgRating   float32  `json:"avgRating"`
+	AvgRating   float64  `json:"avgRating"`
 	ChType      ChType   `json:"chtype"`
 	Description string   `json:"description"`
 	Location    Location `json:"location"`
